+++ packages/@cldcvr/flow-lineage/src/components/f-lineage/draw/draw-lineage.ts
@@ -2,7 +2,7 @@ import { DrawLineageParams } from "../lineage-types";
 import drawLinks from "./draw-links";
 import drawNodes from "./draw-nodes";
 
-export default function drawLineage(params: DrawLineageParams) {
+export default async function drawLineage(params: DrawLineageParams) {
   drawLinks(params);
   drawNodes(params);
 }